+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -81,6 +81,11 @@ New checks
   Warns if a non-overridden virtual function is declared in final class and
   warns if an overridden function is not marked final in a final class.
 
+- New :doc:`bsl-friend-decl
+  <clang-tidy/checks/bsl-friend-decl>` check.
+
+  Warns on friend declarations.
+
 - New :doc:`bsl-function-name-use
   <clang-tidy/checks/bsl-function-name-use>` check.